Moving merge_outputs again to prevent a loop
Having merge_outputs in global utils can result in import loops
in rather unexpected places. Moving it to a different scope to
prevent those from appearing.
